client/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
coverage==7.1.0
idna==3.4
importlib-metadata==4.13.0
markdown-it-py==2.1.0
mdurl==0.1.2
nanoid==2.0.0
Pygments==2.14.0
PyJWT==2.6.0
requests==2.28.2
rich==13.3.1
textual==0.10.1
urllib3==1.26.14
websockets==10.4
zipp==3.12.0
